refactor: machine 관련 로직 추상화
